refactor: isolate shell bootstrap workflow
